refactor(ramses-v3): index pool FeeAdjustment as a pool event
The pool emits FeeAdjustment(oldFee, newFee) on setFee. Decode it in
map_events and apply the fee update in event_to_attributes_updates,
instead of separately listening for the factory's FeeAdjustment in
map_pools_created.
