+++ protocols/substreams/polygon-ramses-v3/src/abi/factory.rs
@@ -9,95 +9,6 @@ pub mod functions {
 pub mod events {
     use super::INTERNAL_ERR;
     #[derive(Debug, Clone, PartialEq)]
-    pub struct FeeAdjustment {
-        pub pool: Vec<u8>,
-        pub new_fee: substreams::scalar::BigInt,
-    }
-    impl FeeAdjustment {
-        const TOPIC_ID: [u8; 32] = [
-            228u8,
-            172u8,
-            203u8,
-            174u8,
-            232u8,
-            47u8,
-            184u8,
-            51u8,
-            172u8,
-            32u8,
-            125u8,
-            76u8,
-            68u8,
-            84u8,
-            197u8,
-            160u8,
-            78u8,
-            133u8,
-            245u8,
-            225u8,
-            233u8,
-            162u8,
-            10u8,
-            158u8,
-            44u8,
-            152u8,
-            229u8,
-            78u8,
-            135u8,
-            6u8,
-            255u8,
-            43u8,
-        ];
-        pub fn match_log(log: &substreams_ethereum::pb::eth::v2::Log) -> bool {
-            if log.topics.len() != 1usize {
-                return false;
-            }
-            if log.data.len() != 64usize {
-                return false;
-            }
-            return log.topics.get(0).expect("bounds already checked").as_ref()
-                == Self::TOPIC_ID;
-        }
-        pub fn decode(
-            log: &substreams_ethereum::pb::eth::v2::Log,
-        ) -> Result<Self, String> {
-            let mut values = ethabi::decode(
-                    &[ethabi::ParamType::Address, ethabi::ParamType::Uint(24usize)],
-                    log.data.as_ref(),
-                )
-                .map_err(|e| format!("unable to decode log.data: {:?}", e))?;
-            values.reverse();
-            Ok(Self {
-                pool: values
-                    .pop()
-                    .expect(INTERNAL_ERR)
-                    .into_address()
-                    .expect(INTERNAL_ERR)
-                    .as_bytes()
-                    .to_vec(),
-                new_fee: {
-                    let mut v = [0 as u8; 32];
-                    values
-                        .pop()
-                        .expect(INTERNAL_ERR)
-                        .into_uint()
-                        .expect(INTERNAL_ERR)
-                        .to_big_endian(v.as_mut_slice());
-                    substreams::scalar::BigInt::from_unsigned_bytes_be(&v)
-                },
-            })
-        }
-    }
-    impl substreams_ethereum::Event for FeeAdjustment {
-        const NAME: &'static str = "FeeAdjustment";
-        fn match_log(log: &substreams_ethereum::pb::eth::v2::Log) -> bool {
-            Self::match_log(log)
-        }
-        fn decode(log: &substreams_ethereum::pb::eth::v2::Log) -> Result<Self, String> {
-            Self::decode(log)
-        }
-    }
-    #[derive(Debug, Clone, PartialEq)]
     pub struct PoolCreated {
         pub token0: Vec<u8>,
         pub token1: Vec<u8>,
