+++ protocols/substreams/polygon-ramses-v3/src/pb/.last_generated_hash
@@ -1,1 +1,1 @@
-a3ba0f236c018b9da0309c64f5c16d74eab631d850a4eb0d3fd67522198f396f
+f6192d5966ff5d7c03271c543f87fc880cbd8192bc93032333f74d08fbb84882
